requirements.txt
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
pypdf2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
pdf2image>=1.16.0
pytesseract>=0.3.10
caj2pdf>=0.1.0
rispy>=0.7.0
pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
rank-bm25>=0.2.0
scikit-learn>=1.3.0
python-pptx>=0.6.21
pdfplumber>=0.9.0
huggingface-hub>=0.16.0